Stop playing the videos when user clicks on the Share icon from the  Edit screen. Currently, it’s continuing to play in the background.
Segment total time is dynamically updating while playing. It
should remain the same and only the current time should run.

Remove the '0' when seconds is single digit

Edit Screen - Done
Saves the segment, but it doesn't go to the Camera screen
like it's supposed to.

On iPhone, it drags the position marker to the start of the segment.
The requirement is to drag the position marker and place it 1 second
before the right marker when released.